Moving files so that all scripts are in the same place
